BUG/MINOR: fixed 'struct buffer' 'size' member calculation
The value of that member of the structure is only used within the
buffer_grow() function, so this does not affect the rest of the program.
Admittedly, the value of that function depended on the value of 'size'
member, but it was never used in this form anyway (just checking whether
the function was executed or not).

Also, checking that the value of parameter 'n' is 0 is added to the function.

Version of the program changed to v1.2.11.