BUG/MINOR: multiple execution of the function worker_stop() disabled
In case HAProxy is used in master-worker mode, it is possible that it sends
a SIGINT signal several times in a row.  Also, it may happen that the program
runs out of time and immediately after that the program receives a SIGINT
signal.  Any of these causes led to the crash of the program.

Version of the program changed to v1.2.12.