MINOR: debug: show return values of all functions in the debug output
If the program is configured using the --enable-debug option, then log
messages are printed to stdout during operation (or to a file specified
at program startup).  In the log one can then find (among other things)
the order in which the function is called and the value that the function
returns (if it is not a void type).

Prior to applying this patch, no value returned by a function was logged.

Log output example:
  [ 1][    0.000661] worker_thread(0x558c674ba200) {
  [ 1][    0.000672]    Worker started, thread id: 104866
  [ 1][    0.000716]    ev_backend_type(0x7f7858000b60) {
  [ 1][    0.000721]       ev_backend_name(4) {
  [ 1][    0.000722]       } = 0x558c6554b05c
  [ 1][    0.000722]    } = 0x558c6554b05c
  [ 1][    0.000723]    libev: using backend 'epoll'
  ..
  [ 1][    5.210447]    worker_thread_exit(0x558c674ba200) {
  [ 1][    5.210458]       Worker is stopped
  [ 1][    5.210459]    } = (nil)

Version of the program changed to v1.2.14.