BUG/MINOR: deferred calling the libev socket timer callback function
On Linux systems with a newer cURL library (say ubuntu 22 which uses cURL
v7.81.0), the program crashes in the mir_curl_timer_cb() function (when
calling function mir_curl_ev_timer_cb()).

It was solved so that the call of the respective function is done via the
timer (and not directly as before).

This should fix GitHub issue #31.

Version of the program changed to v1.2.16.