MINOR: util: added possibility of decoding SPOP frames
In the decode-data utility, the possibility of decoding SPOP frames has
been added.  If we want to display the data in that form, the '-f' option
is used when starting the decode-data program.  The old way of displaying
data (as strings/one-byte data) is achieved by using the option '-s'.

./decode-data -h

Usage: decode-data { -h --help }
       decode-data { -V --version }
       decode-data [OPTION]...

Options are:
  -f data  display data as decoded frame.
  -s data  Display data as strings.
  -h       Show this text.
  -V       Show program version.

Copyright 2023 HAProxy Technologies
SPDX-License-Identifier: GPL-2.0-or-later