MINOR: added possibility to change HTTP headers that are sent to mirror URL
The function spoa_msg_arg_hdrs() has been added, which allows changing the
type of the SPOE argument 'arg_hdrs' to a string (until now, that argument
could only be binary).  In this way, the use of the regsub() converter is
enabled in that argument, so that it is now possible to change the HTTP
headers that are sent to the mirror URL.

For example, instead of:

  args arg_method=method arg_path=url arg_ver=req.ver arg_hdrs=req.hdrs_bin arg_body=req.body

, the following can be used:

  args arg_method=method arg_path=url arg_ver=req.ver arg_hdrs=req.hdrs,regsub(foo,bar) arg_body=req.body

Version of the program changed to v1.2.18.