BUG/MINOR: added support for absolute-form request target
When a request is made to the proxy, the client must send the target URL
in absolute-form as the request-target.  In that case, the creation of
the HTTP request was not performed correctly and the URL to which the
program connects was incorrect.

For example, the address 'http://localhost:8100/', to which the HTTP
request 'GET http://1.2.3.4:12345/index.html HTTP/1.1' was mirrored,
ended up looking like this:

  http://localhost:8100http://1.2.3.4:12345/index.html

Script 'test/get-absolute-uri.sh' has been added for easier testing of
the described problem.

WARNING:  If the cURL version is less than 7.55.0, the CURLOPT_REQUEST_TARGET
option is not used when configuring curl_easy_setopt() handle.  It also
means that sending the target URL in absolute-form as request-target
will not work properly.