BUG/MINOR: fixed the use of '--mirror-(interface|local-port)' program command-line arguments
The ability to set several parameters for outgoing connections was added
to the program a long time ago: interface/address and local port range.

Indeed?  Everything was written, parsing the arguments and setting the
parameters for the output connections, but.. the link between them was
missing.  This commit fixed it, finally.

Version of the program changed to v1.2.19.