MINOR: added '--connection-timeout' and '--timeout' program command-line arguments
The first option sets the maximum time that is allowed in order to
connect to the mirror server.  This was previously hard-coded to
20 seconds.

The second option sets the maximum duration of the single transfer
operation (HTTP requests) with the mirror server.  It was also
hard-coded before, to 10 seconds.

For both settings, the default is set to the previous hard-coded value.