BUILD/MINOR: added support for separate build directory
When building a program, it is possible to create a separate directory
for configuring and compiling the program.  This is useful because in
this way it is possible to make a separate build for release and debug.

The procedure can be as follows:

  % ./scripts/bootstrap
  % mkdir _debug
  % cd _debug
  % ../configure --prefix=/opt --enable-debug
  % make

  % cd ..

  % mkdir _release
  % cd _release
  % ../configure --prefix=/opt
  % make

The possibility of building a debug version of the program (which
adequately has the extension _dbg) was also added.